+++ src/layers/polygon.ts
@@ -50,12 +50,16 @@ export default function getPolygonLayer(
       .value as GradientBinningMethod,
     classCount: settings.lineGradient.classCount.value,
     definedInterval: settings.lineGradient.definedInterval.value,
+    manualBreaks: settings.lineGradient.manualBreaks.value,
+    manualColors: settings.lineGradient.manualColors.value,
   };
   const fillGradientSettings = {
     method: settings.fillGradient.binningMethod.value
       .value as GradientBinningMethod,
     classCount: settings.fillGradient.classCount.value,
     definedInterval: settings.fillGradient.definedInterval.value,
+    manualBreaks: settings.fillGradient.manualBreaks.value,
+    manualColors: settings.fillGradient.manualColors.value,
   };
   const lineColor = createLayerColorAccessor<PolygonFeature>({
     items: data,
@@ -147,6 +151,8 @@ export default function getPolygonLayer(
           settings.lineGradient.binningMethod.value.value,
           settings.lineGradient.classCount.value,
           settings.lineGradient.definedInterval.value,
+          settings.lineGradient.manualBreaks.value,
+          settings.lineGradient.manualColors.value,
           getNumericColorBinsSignature(lineColor.bins),
         ],
         [
@@ -168,6 +174,8 @@ export default function getPolygonLayer(
           settings.fillGradient.binningMethod.value.value,
           settings.fillGradient.classCount.value,
           settings.fillGradient.definedInterval.value,
+          settings.fillGradient.manualBreaks.value,
+          settings.fillGradient.manualColors.value,
           getNumericColorBinsSignature(fillColor.bins),
         ],
         [